requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
xgboost==2.0.3
scikit-learn==1.5.0
imbalanced-learn==0.12.3
shap==0.45.1
pandas==2.2.2
numpy==1.26.4
mlflow==2.13.0
lightgbm
ipykernel
seaborn
streamlit